+++ src/main/java/net/dc/jademc/item/armor/EnderiteHelmet.java
@@ -1,0 +1,12 @@
+package net.dc.jademc.item.armor;
+
+import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.item.ArmorItem;
+
+public class EnderiteHelmet extends ArmorItem {
+    public EnderiteHelmet() {
+        super(new EnderiteArmorMaterial(), EquipmentSlotType.HEAD, EnderiteArmorMaterial.PROPERTIES);
+
+        this.setRegistryName("enderite_helmet");
+    }
+}