+++ src/main/java/net/dc/jademc/item/armor/EnderiteLeggings.java
@@ -1,0 +1,12 @@
+package net.dc.jademc.item.armor;
+
+import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.item.ArmorItem;
+
+public class EnderiteLeggings extends ArmorItem {
+    public EnderiteLeggings() {
+        super(new EnderiteArmorMaterial(), EquipmentSlotType.LEGS, EnderiteArmorMaterial.PROPERTIES);
+
+        this.setRegistryName("enderite_leggings");
+    }
+}